bookmyshow/requirements.txt
google-adk
dotenv
supabase